Dune files for new package 'http_pirc'; compiling http projections.
Updated dune files to create a new package 'http_pirc', allowing for a new way to compile projected
code for the http backend.

This was done due to how the http backend works, which produces projected code that relies on a module
called 'http_pirc'. Before, it was not configured so that 'http_pirc' was treated as its own package,
as listed in the 'dune-project' file. This meant that running 'dune build @install' (that is, building
the '@install' target using the 'dune' command) would not produce the appropriate files for use with
the 'ocamlfind' command when used later to with the 'ocamlopt' subcommand to compile the projected
programs.

The changes to the dune files now allow the use of the 'ocamlfind' command to compile programs
projected for the http backend, for example:
  $ OCAMLPATH=./_build/install/default/lib/ ocamlfind ocamlopt -o test.exe -linkpkg -package http_pirc test.ml
However, this also requires that the './_build/install' directory is generated by dune, which can be
done with the following command:
  $ dune build @install

This method displaces the old method of compiling projected programs for the http backend, which would
programmatically generate a dune file in the 'examples' directory (note: as of this commit, the
compiler still assumes all input files are going to be in the 'examples' directory), which would then
need to be run using 'dune build examples', which would then produce the compiled projected programs
in the '_build/default/examples/' directory.