Fix wrong function used, introduced when merging ELR repo.
When merging the ELR repo, there were many conflicts involved in the ocamlgen/netgen directory due
to a recent restructuring of the code in the FOD repo. In the process of copying a function used to
initialize part of the HTTP "toplevel" code, the function for initializing the domains "toplevel" was
copied instead (both had the same name, but in different scopes). This commit fixes this issue.