Fix Config_parser module not available in HTTP projections.
Since the bug in the prior commit was fixed (where the wrong function was copied), now trying to
compile the projections for the http backend does not work since the projected code relies on the
Config_parser module, which is not part of the http_pirc module. The easiest fix was to make
Config_parser a submodule of the http_pirc library.