Removed HTTP backend dune file generation.
Due to the changes in commit b17c140e35 (which introduced the http_pirc package to make it possible to
compile HTTP projections using ocamlfind/ocamlopt), removed the code to generate a dune file when
using the HTTP backed, which was the old method of compiling projections for the HTTP backend. Also
removed some supporting code related dune file generation, including code that would add an extra
import ('open Ffi_lib') to the HTTP projections.