Remove hardcoded 'examples' dir when compiling a '.pir' prog.
The pirc compiler used to just use the basename of the input Pirouette file and would assume it was
always located in a directory called 'examples', and would always put the output .ml file(s) in the
examples directory. This commit fixes it so that the input directory can be from anywhere and that the
output .ml file(s) will be placed in the same directory as the input file.