Fixup Ian's automated testing pipeline and changes to pirproj.sh.
Made some changes to 'test/dune' and test/pipeline_gen.ml to allow Ian's testing pipeline to work
correctly.

Two new flags were added to pirproj.sh to support using Ian's testing pipeline (due some
limitations of dune that I ran into).

The test programs and '.ans' files in 'test/pipeline_src' were also changed. This was because if
multiple processes output things, their order of outputs might not be the same each time. Thus, the
testing pipeline was modified to do a stable sort of the output (on both the .res and .ans files),
sorting just the first "field" of each line, where the first field is the process name followed by
colon. Doing this means that all the messages w.r.t. a single process are ordered chronologically, but
the ordering of messages between processes is not considered.