Update Compilation, FFI, ast-dump, and pirc CLI.
The pirc executable now compiles the projected ml files for the user. If the old
behavior is desired, the '-projections-only' flag may be used to have pirc skip
performing compilation on the ml files.

FFI changes:
Instead of the old FFI syntax:
  `foreign myFunc : unit -> unit := "@module:func";`
New syntax is introduced for FFI:
  `foreign myFunc : unit -> unit := "package:submodule.func@path";`
  `foreign myFunc : unit -> unit := "module.func@path";`
where in either case, the '@path' part is optional.
The 'package' part specifies which packages to include when compiling the
ml files, and is also the topmost module that the function 'func' belongs
to. The packages used by FFI are automatically searched for and included by
ocamlfind when pirc compiles the projected ml files.
The 'submodule' part is one or more nested submodules the function 'func' is
from (with periods between submodule names).
If 'package' is not specified, then the 'module' part is like the
'submodule' part, except the topmost module is also specified. A FFI binding
specified in this way must either be part of Ocaml's standard library or be
from a module in an ml file included with the new '-ml-files' flag.

ast-dump changes:
An ast-dump is no longer automatically produced everytime pirc is used and will
only happen when the '-ast-dump' flag is used.

CLI changes, new flags:
[1] -package
[2] -projections-only
[3] -ml-files
[4] -w